pyproject.toml
[project]
name = "e-commerce-recommendation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ravikumarchavva", email = "ravikumarchavva@outlook.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "googletrans>=4.0.2",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "kagglehub[pandas-datasets]>=0.4.0",
    "openai>=2.15.0",
    "pillow-simd>=9.5.0.post2",
    "polars>=1.37.1",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.0",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.46",
    "torch>=2.9.1",
    "torch-c-dlpack-ext>=0.1.5",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.6",
    "vllm>=0.14.0",
]

[project.scripts]
e-commerce-recommendation = "e_commerce_recommendation:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
